requirements.txt
# requirements.txt
streamlit>=1.28.0
streamlit-chat
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-google-genai
langchain-cohere
langchain-openai # For OpenRouter, Together, Groq
langchain-huggingface
requests>=2.31.0 # For Ollama, etc.
keyring # For secure API key storage
pypdf # PDF processing
PyPDF2>=3.0.0
faiss-cpu # Vector store for PDF search
python-dotenv>=1.0.0 # Optional, for .env file support
tavily-python>=0.3.0 # <<< ADDED: Tavily client
langchainhub # <<< ADDED: To pull agent prompts easily
tiktoken>=0.5.1
st-copy-to-clipboard>=0.0.4
python-docx>=1.0.1
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
chromadb>=0.4.0
ninja>=1.11.0  # For NVIDIA NIM provider

# Potentially needed OS-specific keyring backends:
# keyring[linux] # Example for Linux with dbus
# keyring[windows] # Example for Windows